tasks/src/models/task.ts
import mongoose from 'mongoose';
import { updateIfCurrentPlugin } from 'mongoose-update-if-current';

interface TaskAttrs {
  projectId: string;
  description: string;
  completed?: boolean;
}

interface TaskDoc extends mongoose.Document {
  projectId: string;
  description: string;
  completed: boolean;
  version: number;
}

interface TaskModel extends mongoose.Model<TaskDoc> {
  build(attrs: TaskAttrs): TaskDoc;
}

const taskSchema = new mongoose.Schema(
  {
    projectId: {
      type: mongoose.Schema.Types.ObjectId,
      required: true
    },
    description: {
      type: String,
      required: true
    },
    completed: {
      type: Boolean,
      required: true,
      default: false
    }
  },
  {
    toJSON: {
      transform(doc, ret) {
        ret.id = ret._id;
        delete ret._id;
      }
    }
  }
);
taskSchema.set('versionKey', 'version');
taskSchema.plugin(updateIfCurrentPlugin);

taskSchema.statics.build = (attrs: TaskAttrs) => {
  return new Task(attrs);
};

const Task = mongoose.model<TaskDoc, TaskModel>('Task', taskSchema);

export { Task };
